+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add a way to scale the UI in the settings
 - Display RGB slider values in 0-255 range as integers rather than floats
 - Move harmonies box outside of the harmonies header
+- Add a way to hide the harmonies color box
 
 # 0.9.0
 - Change button layout in palette view